library_manangement/registration/urls.py
from django.urls import path 
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)
from .views import RegisterView, LogoutView
urlpatterns = [

    path('api/login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/login/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/register',RegisterView.as_view(), name="registration view"),
    path('api/logout/', LogoutView.as_view(), name='auth_logout'),

]