requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
